# dist/tsoa/swagger.json
{
    "components": {
...
        "schemas": {
...
            "ProductRequestParameter": {
                "properties": {
                    "stok": {
                        "type": "number",
                        "format": "double"
                    },
                    "hargaJual": {
                        "type": "number",
                        "format": "double"
                    },
                    "hargaModal": {
                        "type": "number",
                        "format": "double"
                    },
                    "tanggalMasuk": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "vendorId": {
                        "type": "number",
                        "format": "double"
                    },
                    "name": {
                        "type": "string"
                    },
                    "sku": {
                        "type": "string"
                    }
                },
                "required": [
                    "stok",
                    "hargaJual",
                    "hargaModal",
                    "tanggalMasuk",
                    "vendorId",
                    "name",
                    "sku"
                ],
                "type": "object"
            },
            "StockRequestParameter": {
                "properties": {
                    "vendorId": {
                        "type": "number",
                        "format": "double"
                    },
                    "productId": {
                        "type": "number",
                        "format": "double"
                    },
                    "sell_price": {
                        "type": "number",
                        "format": "double"
                    },
                    "buy_price": {
                        "type": "number",
                        "format": "double"
                    },
                    "total_stock": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "type": "object"
            },
            "TransactionRequestParameter": {
                "properties": {
                    "detail": {
                        "items": {
                            "properties": {
                                "sub_total": {
                                    "type": "number",
                                    "format": "double"
                                },
                                "productId": {
                                    "type": "number",
                                    "format": "double"
                                },
                                "final_price": {
                                    "type": "number",
                                    "format": "double"
                                },
                                "amount": {
                                    "type": "number",
                                    "format": "double"
                                }
                            },
                            "required": [
                                "sub_total",
                                "productId",
                                "final_price",
                                "amount"
                            ],
                            "type": "object"
                        },
                        "type": "array"
                    },
                    "amount_paid": {
                        "type": "number",
                        "format": "double"
                    },
                    "customer_id": {
                        "type": "number",
                        "format": "double"
                    },
                    "actual_total_price": {
                        "type": "number",
                        "format": "double"
                    },
                    "expected_total_price": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "detail",
                    "amount_paid",
                    "customer_id",
                    "actual_total_price",
                    "expected_total_price"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "access_token",
                "in": "header",
                "flow": "implicit",
                "scopes": {
                    "write:pets": "modify things",
                    "read:pets": "read things"
                }
            }
        }
    },
    "info": {
        "title": ""
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/api/products": {
            "get": {
                "operationId": "GetAllProducts",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "api_key": [
                            "read:product"
                        ]
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "CreateProduct",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "api_key": [
                            "create:product"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequestParameter"
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "UpdateProduct",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "api_key": [
                            "update:product"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequestParameter"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteProduct",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "api_key": [
                            "delete:product"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/products/search": {
            "get": {
                "operationId": "SearchProduct",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "api_key": [
                            "read:product"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
        "/api/stock": {
            "get": {
                "operationId": "GetAllStock",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Stock"
                ],
                "security": [
                    {
                        "api_key": [
                            "read:stock"
                        ]
                    }
                ],
                "parameters": []
            }
        },
        "/api/stock/{id}": {
            "put": {
                "operationId": "UpdateStock",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Stock"
                ],
                "security": [
                    {
                        "api_key": [
                            "update:stock"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StockRequestParameter"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteStock",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Stock"
                ],
                "security": [
                    {
                        "api_key": [
                            "delete:stock"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "patch": {
                "operationId": "PatchStock",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Stock"
                ],
                "security": [
                    {
                        "api_key": [
                            "update:stock"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StockRequestParameter"
                            }
                        }
                    }
                }
            }
        },
        "/api/stock/search/{query}": {
            "get": {
                "operationId": "FindStock",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Stock"
                ],
                "security": [
                    {
                        "api_key": [
                            "read:stock"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/transaction": {
            "get": {
                "operationId": "GetAllTransaction",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "api_key": [
                            "read:transaction"
                        ]
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "CreateTransaction",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "api_key": [
                            "create:transaction"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionRequestParameter"
                            }
                        }
                    }
                }
            }
        },
        "/api/transaction/{id}": {
            "put": {
                "operationId": "UpdateTransaction",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "api_key": [
                            "update:transaction"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionRequestParameter"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteTransaction",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "api_key": [
                            "delete:transaction"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/transaction/search/{query}": {
            "get": {
                "operationId": "SearchTransaction",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "api_key": [
                            "read:transaction"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
}